+++ tests/test_align_min_metadata.py
@@ -10,8 +10,8 @@ import sys
 from valis import registration
 
 
-from valis.slide_tools import *
-from valis.slide_io import *
+# from valis.slide_tools import *
+# from valis.slide_io import *
 
 def get_parent_dir():
     cwd = os.getcwd()
@@ -36,16 +36,8 @@ def test_align_min_metadata():
     registrar = registration.Valis(src_dir, dst_dir)
     registrar.register()
 
-    src_f = registrar.original_img_list[0]
-    self = VipsSlideReader(src_f)
-
     registrar.warp_and_save_slides(dst_dir)
-    # self = registrar
     registrar.warp_and_merge_slides(dst_f=os.path.join(dst_dir, "merged.ome.tiff"))
-
-    from valis.slide_io import *
-    self = slide_obj.reader
-    src_f =
 
     assert True